rename: components -> scripts로 변경
