+++ apps/web/app/layout.tsx
@@ -3,7 +3,7 @@ import { RootLayout as ViewportLayout } from '@sambad/web-domains/commmon';
 import dayjs from 'dayjs';
 import localFont from 'next/font/local';
 
-import { KakaoScript } from '../components/KakaoScript';
+import { KakaoScript } from '../scripts/KakaoScript';
 
 import type { Metadata } from 'next';
